Fixed issue reported by Uri. Refuse to download null named included files
