+++ lib/composer.js
@@ -141,6 +141,9 @@
         tag = this.resolve(nodes.ScalarNode, event.value, event.implicit);
       }
       if (event.tag === '!include') {
+        if (event.value.match(/^\s*$/)) {
+          throw new exports.ComposerError('while composing scalar out of !include', null, "file name/URL cannot be null", event.start_mark);
+        }
         if (this.src != null) {
           event.value = require('url').resolve(this.src, event.value);
         }
