+++ lib/nodes.js
@@ -232,9 +232,13 @@
         });
         if (node_has_property) {
           return _this.value.forEach(function(ownNodeProperty) {
-            var ownNodePropertyName;
+            var nonNullNode, ownNodePropertyName;
             ownNodePropertyName = ownNodeProperty[0].value;
             if ((ownNodePropertyName === name) || ((ownNodePropertyName + '?') === name) || (ownNodePropertyName === (name + '?'))) {
+              if ((ownNodeProperty[1].tag === "tag:yaml.org,2002:null") && (resourceProperty[1].tag === "tag:yaml.org,2002:map")) {
+                nonNullNode = new exports.MappingNode('tag:yaml.org,2002:map', [], ownNodeProperty[1].start_mark, ownNodeProperty[1].end_mark);
+                ownNodeProperty[1] = nonNullNode;
+              }
               ownNodeProperty[1].combine(resourceProperty[1]);
               return ownNodeProperty[0].value = ownNodeProperty[0].value.replace(/\?$/, '');
             }
