+++ test/parser.html
@@ -14,6 +14,7 @@
       mocha.reporter('html');
       expect = chai.expect;
     </script>
+    <script src="regressions.js"></script>
     <script src="parser.js"></script>
     <script src="resourceTypesValidations.js"></script>
     <script src="duplicateKeysValidations.js"></script>
